requirements.txt
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
